% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-group-id.R
\name{add_group_id}
\alias{add_group_id}
\title{Helper function to add group ID information}
\usage{
add_group_id(.data, group)
}
\arguments{
\item{.data}{data.frame}

\item{group}{group to add the ID for}
}
\value{
\code{data.frame} with additional column named ".group_id", where "group" is
named after the variable you would like to add the group ID for.
}
\description{
Some modelling software needs group data as an index. This means instead of
\code{c("AU", "AU", "UK", "UK", "NZ")} you need \code{c(1, 1, 2, 2, 3)},
corresponding to each of the groups. This function makes it simpler to add
that information to a dataframe.
}
\examples{
malaria_africa_ts \%>\% add_group_id(country)
}
