% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-hts-inla.R
\name{as_hts_inla}
\alias{as_hts_inla}
\title{Create object of class "hts_inla"}
\usage{
as_hts_inla(x, formula)
}
\arguments{
\item{x}{model from inlabru}

\item{formula}{formula passed to \code{fit_hts()}}
}
\value{
hts_inla model
}
\description{
Create object of class "hts_inla"
}
\note{
internal
}
\author{
Nicholas Tierney
}
