% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.hts_inla}
\alias{augment.hts_inla}
\title{augment: Add predictions and residuals to data}
\usage{
\method{augment}{hts_inla}(x, newdata = NULL, exponentiate = FALSE, ...)
}
\arguments{
\item{x}{object of class "hts_inla"}

\item{newdata}{new data to pass to prediction? Default is NULL}

\item{exponentiate}{default FALSE. Whether to exponentiate predictions}

\item{...}{extra arguments to pass to \code{augment()}. Currently unused.}
}
\value{
dataframe with column names of original data, as well as extra columns,  ".fitted" and ".resid".
}
\description{
The \code{augment()} function returns the data used in the model
}
\examples{
hts_example_model
augment(hts_example_model)
}
