% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_likelihoods.R
\name{available_likelihoods}
\alias{available_likelihoods}
\title{Helper function to show the available likelihoods that can be used}
\usage{
available_likelihoods()
}
\value{
list of the available likelihoods
}
\description{
Helper function to show the available likelihoods that can be used
}
\examples{
available_likelihoods()
}
