% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_bru_formula.R
\name{build_bru_formula}
\alias{build_bru_formula}
\title{create formula to pass to inlabru::bru}
\usage{
build_bru_formula(formula, .data, special_index)
}
\arguments{
\item{formula}{formula with an hts() term}

\item{.data}{tsibble data}

\item{special_index}{special index term}
}
\value{
formula
}
\description{
create formula to pass to inlabru::bru
}
\examples{
build_bru_formula(
  formula = pr ~ avg_lower_age + hts(who_subregion, country),
  .data = malaria_africa_ts,
  special_index = month_num
)
}
\author{
Nicholas Tierney
}
