% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hts_data.R
\name{create_hts_data}
\alias{create_hts_data}
\title{Create data to be used when fitting a model with fit_hts}
\usage{
create_hts_data(.data, formula)
}
\arguments{
\item{.data}{tsibble}

\item{formula}{formula with hts() term}
}
\value{
data.frame
}
\description{
Create data to be used when fitting a model with fit_hts
}
\examples{
data_w_groups <- create_hts_data(
  .data = malaria_africa_ts,
  formula = pr ~ avg_lower_age + hts(who_subregion, country)
  )
data_w_groups
}
