% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts-fit.R
\name{fit_hts}
\alias{fit_hts}
\title{Fit a hierarchical time series model}
\usage{
fit_hts(
  formula,
  .data,
  special_index,
  family,
  verbose = TRUE,
  keep_data = TRUE,
  ...
)
}
\arguments{
\item{formula}{formula must contain fixed effect terms and \code{hts()}}

\item{.data}{a \code{tsibble}}

\item{special_index}{The variable that represents time in your data}

\item{family}{character.}

\item{verbose}{default TRUE, whether to display message}

\item{keep_data}{attach the data used to fit the model to the final
model object?}

\item{...}{list of options to pass to \code{inlabru::bru}, see \code{?inlabru::bru}
to see these options.}
}
\value{
inlabru model
}
\description{
Fit a hierarchical time series model
}
\section{How to allow weighting of observations?}{
To weight your observations, say when using empirical logit, you must
provide a vector of variances, for each observation, which represents
the uncertainty of measurement. You can do this, provided the
\code{family = "gaussian"}, add the argument, \code{scale = vec}, to \code{fit_hts},
where \code{vec} is a vector of variances for each observation.
}

\examples{
\dontrun{
m <- fit_hts(
#inputs are  the levels of hierarchy, in order of decreasing size
formula = pr ~ avg_lower_age + hts(who_subregion, country),
.data = malaria_africa_ts,
family = "gaussian",
special_index = month_num
)
}
}
