% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.hts_inla}
\alias{fitted.hts_inla}
\title{Extract fitted values from model object}
\usage{
\method{fitted}{hts_inla}(object, fit = "mean", exponentiate = FALSE, ...)
}
\arguments{
\item{object}{hts_inla object}

\item{fit}{default is "mean". Options are: "mean", "sd", "0.025quant", "0.5quant", "0.975quant", and "mode".}

\item{exponentiate}{default FALSE}

\item{...}{extra arguments. Currently not used}
}
\description{
Extract fitted values from model object
}
