% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_vars_in_data.R
\name{formula_vars_in_data}
\alias{formula_vars_in_data}
\title{Check if formula variables are present in data}
\usage{
formula_vars_in_data(data, formula)
}
\arguments{
\item{data}{a \code{tsibble}}

\item{formula}{formula}
}
\description{
Check if formula variables are present in data
}
\examples{

formula_vars_in_data(
data = malaria_africa_ts,
formula = pr ~ dates
)

}
