% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.R
\name{glance.hts_inla}
\alias{glance.hts_inla}
\title{Glance: A single row summary of a model fit}
\usage{
\method{glance}{hts_inla}(x, ...)
}
\arguments{
\item{x}{object of class "hts_inla"}

\item{...}{extra arguments to pass to \code{glance()}. Currently unused}
}
\value{
dataframe with column names: deviance information criterion, wakaike information criteriot, log marginal-likelihood (integration), log marginal-likelihood (Gaussian).
}
\description{
Glance: A single row summary of a model fit
}
\examples{
hts_example_model
glance(hts_example_model)
}
