% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts-builder.R
\name{hts_builder}
\alias{hts_builder}
\title{Build hts call}
\usage{
hts_builder(.data, ...)
}
\arguments{
\item{.data}{tsibble}

\item{...}{the levels of hierarchy, in order of decreasing size}
}
\value{
list of expressions to pass into INLA
}
\description{
Build hts call
}
\examples{

hts_builder(brolgar::wages, who_region, who_subregion, country)

hts_builder(who_region, who_subregion, country) \%>\%
  paste0(collapse = " + ") \%>\%
  rlang::sym()

}
