% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-hts-example-model.R
\docType{data}
\name{hts_example_model}
\alias{hts_example_model}
\title{Example model for use in testing and examples}
\format{
An object of class \code{hts_inla} (inherits from \code{hts_inla}, \code{list}) of length 8.
}
\usage{
data(hts_example_model)
}
\description{
Example model for use in testing and examples
}
\keyword{datasets}
