% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_yahstee_deps.R
\name{install_yahtsee_deps}
\alias{install_yahtsee_deps}
\title{Install yahtsee dependencies}
\usage{
install_yahtsee_deps()
}
\value{
nothing
}
\description{
Currently this just installs the latest version of INLA
}
\examples{
\dontrun{
install_yahtsee_deps()
}
}
