% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-malaria-africa-ts.R
\docType{data}
\name{malaria_africa_ts}
\alias{malaria_africa_ts}
\title{Prevalence Rate data of Malaria in Africa}
\format{
An object of class \code{tbl_ts} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1046 rows and 15 columns.
}
\usage{
data(malaria_africa_ts)
}
\description{
Data containing information on malaria prevalence in Africa. Data was
extracted using the \code{malariaAtlas} package. If using this data, please
cite:
}
\details{
Pfeffer, D.A., Lucas, T.C., May, D., Harris, J., Rozier, J., Twohig, K.A.,
Dalrymple, U., Guerra, C.A., Moyes, C.L., Thorn, M., Nguyen, M., et al.
2018. malariaAtlas: an R interface to global malariometric data hosted by
the Malaria Atlas Project. Malaria Journal, 17(1), p.352.

Format: a 'tsibble' data frame with 1,117 observations on the following 15
variables.
\describe{
\item{\code{who_region}}{A character of WHO region, the largest region size.}
\item{\code{who_subregion}}{A character of WHO subregion, the second largest
region size.}
\item{\code{country}}{A character of countries, the smallest region size.}
\item{\code{country_id}}{A three letter character code adhering to the
standard, ISO 3166-1 alpha-3. For more detail see
\url{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3.}}
\item{\code{continent_id}}{A character of continent name}
\item{\code{year}}{integer of year data recorded}
\item{\code{month}}{integer of month data recorded}
\item{\code{date}}{date class of year+month recorded. Defaults to day = 1 as day was not provided.}
\item{\code{lower_age}}{lower age range}
\item{\code{upper_age}}{upper age range}
\item{\code{examined}}{integer, number of people examined}
\item{\code{positive}}{numeric, number of positive cases}
\item{\code{pr}}{positive rate - given from positive/examined}
\item{\code{species}}{Mosquito species}
\item{\code{method}}{method of measurement}
}
}
\keyword{datasets}
