% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.hts_inla}
\alias{predict.hts_inla}
\title{Predict from hts_inla model}
\usage{
\method{predict}{hts_inla}(
  object,
  newdata = NULL,
  type = c("link", "response", "terms"),
  se.fit = FALSE,
  dispersion = NULL,
  terms = NULL,
  na.action = na.pass,
  ...
)
}
\arguments{
\item{object}{a fitted object of class inheriting from \code{hts_model}}

\item{newdata}{optionally, a data frame in which to look for variables with
which to predict.  If omitted, the fitted linear predictors are used.}

\item{type}{the type of prediction required.  The default is on the scale of
the linear predictors; the alternative ‘"response"’ is on the scale of
the response variable.  Thus for a default binomial model the default
predictions are of log-odds (probabilities on logit scale) and
‘type = "response"’ gives the predicted probabilities.  The ‘"terms"’
option returns a matrix giving the fitted values of each term in the
model formula on the linear predictor scale. The value of this argument
can be abbreviated.}

\item{se.fit}{logical switch indicating if standard errors are required.}

\item{dispersion}{the dispersion of the GLM fit to be assumed in computing
the standard errors.  If omitted, that returned by ‘summary’ applied to
the object is used.}

\item{terms}{with ‘type = "terms"’ by default all terms are returned.  A
character vector specifies which terms are to be returned}

\item{na.action}{function determining what should be done with missing
values in ‘newdata’.  The default is to predict ‘NA’.}

\item{...further}{arguments passed to or from other methods.}
}
\value{

}
\description{
Predict from hts_inla model
}
\note{
currently borrowing parameters + descriptions from predict.glm
}
\examples{
\dontrun{
predict(hts_example_model)
}
}
