% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-if-terms-repeated.R
\name{test_if_terms_repeated}
\alias{test_if_terms_repeated}
\title{.. content for \description{} (no empty lines) ..}
\usage{
test_if_terms_repeated(formula)
}
\description{
.. content for \description{} (no empty lines) ..
}
