% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.hts_inla}
\alias{tidy.hts_inla}
\title{Tidy:  summarise a model's statistical findings}
\usage{
\method{tidy}{hts_inla}(x, ...)
}
\arguments{
\item{x}{object of class "hts_inla"}

\item{...}{extra arguments to pass to \code{tidy()}. Currently unused}
}
\value{
dataframe with column names: "variable", "mean", "sd", and quantiles for 0.025, 0.5, and 0.975, as well as mode and kld
}
\description{
The \code{tidy()} function provides a data frame of the mean, standard deviation,
quantiles
}
\examples{
hts_example_model
tidy(hts_example_model)
}
