% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-who-regions.R
\docType{data}
\name{who_regions}
\alias{who_regions}
\title{Who Regions}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 110 rows and 4 columns.
}
\usage{
data(who_regions)
}
\description{
Data containing information on countries and WHO designated regions and
ISO 3166-1 alpha-3 country code. From largest to smallest area it goes:
"who_region", "who_subregion", then "country".
}
\details{
Format: a data frame with 110 observations on the following 4 variables.
\describe{
\item{\code{who_region}}{A character of WHO region, the largest region size.}
\item{\code{who_subregion}}{A character of WHO subregion, the second largest
region size.}
\item{\code{country}}{A character of countries, the smallest region size.}
\item{\code{country_iso_code}}{A three letter character code adhering to the
standard, ISO 3166-1 alpha-3. For more detail see
\url{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3.}}
}
}
\keyword{datasets}
