% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahtsee-package.R
\docType{package}
\name{yahtsee-package}
\alias{yahtsee}
\alias{yahtsee-package}
\title{yahtsee: Hierarchical Time Series Models for Malaria}
\description{
Hierarchical Time Series Models for Malaria.
}
\author{
\strong{Maintainer}: Nicholas Tierney \email{nicholas.tierney@gmail.com} (\href{https://orcid.org/0000-0003-1460-8722}{ORCID})

}
\keyword{internal}
