% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahtsee-sitrep.R
\name{yahtsee_sitrep}
\alias{yahtsee_sitrep}
\title{Checks if system requirements are present}
\usage{
yahtsee_sitrep()
}
\value{
message on availability of system requirements
}
\description{
Currently this just checks for INLA
}
\examples{
\dontrun{
yahtsee_sitrep()
}
}
