% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-hts-model.R
\name{as_hts_model}
\alias{as_hts_model}
\title{Create object of class "hts_model"}
\usage{
as_hts_model(x)
}
\arguments{
\item{x}{model from inlabru}
}
\value{
hts model
}
\description{
Create object of class "hts_model"
}
\note{
internal
}
\author{
Nicholas Tierney
}
