% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-fit-hts.R
\name{fit_hts}
\alias{fit_hts}
\title{Fitted Hierarchical Time Series}
\usage{
fit_hts(formula, .data, family, ...)
}
\arguments{
\item{formula}{a formula}

\item{...}{extra args}
}
\description{
Typically this includes a "Specials" section
}
