% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts-fit.R
\name{fit_hts}
\alias{fit_hts}
\title{Fit a hierarchical time series model}
\usage{
fit_hts(formula, .data, special_index, family, ...)
}
\arguments{
\item{formula}{formula must contain fixed effect terms and \code{hts()}}

\item{.data}{a \code{tsibble}}

\item{special_index}{The variable that represents time in your data}

\item{family}{character}

\item{...}{extra arguments to pass to inlabru}
}
\value{
inlabru model
}
\description{
Fit a hierarchical time series model
}
\examples{
m <- fit_hts(
inputs are  the levels of hierarchy, in order of decreasing size
formula = pr ~ avg_lower_age + hts(who_subregion, country),
.data = malaria_africa_ts,
family = "gaussian",
special_index = month_num
)
}
