% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.hts_inla}
\alias{fitted.hts_inla}
\title{Extract fitted values from \code{hts_inla} model object}
\usage{
\method{fitted}{hts_inla}(object, exponentiate = FALSE, ...)
}
\arguments{
\item{object}{hts_inla object}

\item{exponentiate}{default FALSE}

\item{...}{extra arguments. Currently not used.}
}
\description{
This returns the fitted
}
\examples{
head(fitted(hts_example_model))
}
