% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts-builder.R
\name{hts_builder}
\alias{hts_builder}
\title{Build hts call}
\usage{
hts_builder(.data, ...)
}
\arguments{
\item{.data}{tsibble}

\item{...}{the levels of hierarchy, in order of decreasing size}
}
\value{
constructed expression of ar1 terms to pass to inla
}
\description{
Build hts call
}
\examples{

hts_builder(malaria_africa_ts, who_region, who_subregion, country)

}
