% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahtsee-package.R
\docType{package}
\name{yahtsee-package}
\alias{yahtsee}
\alias{yahtsee-package}
\title{yahtsee: Yet Another Hierachical Time Series Extension and Expansion}
\description{
An opinionated approach to building hierarchical time series models in R using INLA and inlabru.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/njtierney/yahtsee}
  \item Report bugs at \url{https://github.com/njtierney/yahtsee/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Tierney \email{nicholas.tierney@gmail.com} (\href{https://orcid.org/0000-0003-1460-8722}{ORCID})

Other contributors:
\itemize{
  \item Mitchell O'Hara-Wild [contributor]
}

}
\keyword{internal}
